Iogonek and misc fixes

+++ sources/build.sh
@@ -1,9 +1,12 @@
 #!/bin/sh
-#set -e
+set -e
 
 
 echo "Generating Static fonts"
 mkdir -p ../fonts
+mkdir -p ../fonts/otf
+mkdir -p ../fonts/ttf
+mkdir -p ../fonts/vf
 fontmake -m JosefinSans.designspace -i -o ttf --output-dir ../fonts/ttf/
 fontmake -m JosefinSans.designspace -i -o otf --output-dir ../fonts/otf/
 fontmake -m JosefinSans-Italic.designspace -o ttf --output-dir ../fonts/ttf/
@@ -22,11 +25,11 @@ for ttf in $ttfs
 do
 	gftools fix-dsig -f $ttf;
 	#ttfautohint $ttf "$ttf.fix";
-	mv "$ttf.fix" $ttf;
+	#mv "$ttf.fix" $ttf;
 done
 
 vfs=$(ls ../fonts/vf/*.ttf)
-
+echo vfs
 echo "Post processing VFs"
 for vf in $vfs
 do
@@ -34,8 +37,6 @@ do
 	#ttfautohint-vf --stem-width-mode nnn $vf "$vf.fix";
 	#mv "$vf.fix" $vf;
 done
-
-
 
 echo "Fixing VF Meta"
 gftools fix-vf-meta $vfs;
@@ -55,13 +56,19 @@ done
 echo "Fixing Hinting"
 for vf in $vfs
 do
-
 	gftools fix-nonhinting $vf "$vf.fix";
-	mv "$vf.fix" $vf;
+	if [ -f "$vf.fix" ]; then mv "$vf.fix" $vf; fi
 done
+
 for ttf in $ttfs
 do
-	gftools fix-nonhinting $ttf "$vf.fix";
-	mv "$ttf.fix" $ttf;
+	gftools fix-nonhinting $ttf "$ttf.fix";
+	if [ -f "$ttf.fix" ]; then mv "$ttf.fix" $vf; fi
 done
-rm ../fonts/*.ttx ../fonts/*gasp.ttf
+
+rm -f ../fonts/vf/*.ttx
+rm -f ../fonts/ttf/*.ttx
+rm -f ../fonts/vf/*gasp.ttf
+rm -f ../fonts/ttf/*gasp.ttf
+
+echo "Done"
